refactor(react_hooks): split scanner.rs into statements and expressions
